exchange-backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bcrypt==4.1.2
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
flask-marshmallow==1.2.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
keras==3.2.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.21.0
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.2
phonenumbers==8.13.35
protobuf==4.25.3
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.2.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.25
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
